Moved ModVersion to jena-base
